\name{addAction-methods}
\docType{methods}
\alias{addAction-methods}
\alias{addAction,ActionQueue,Action-method}
\alias{addAction}

\title{ Method "addAction"}
\description{
 adds a new object of class \code{"\linkS4class{Action}"} to an action queue.
}

\section{Methods}{
\describe{
  \item{\code{signature(currentActionQ = "ActionQueue", newAction = "Action")}}{ 
      adds a new action to an action queue.  }
}}

\keyword{methods}

