\name{DesignSpecifier-class}
\Rdversion{1.1}
\docType{class}
\alias{DesignSpecifier-class}

\title{Class Union "DesignSpecifier"  }
\description{
   An S4 class union, whose member classes (special subclasses) represent specifications for 
   different types of designs.  }
   
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\seealso{
    \code{\link{sim1CT}},
    \code{\linkS4class{APlusBSpecifier}},
    \code{\linkS4class{CRMSpecifier}}
}

\examples{
showClass("DesignSpecifier")
}

\keyword{classes}
